generating strange error

+++ lab5/graphics.c
@@ -128,15 +128,14 @@ int (set_graphics_mode)(uint16_t mode) {
 
 int (vg_draw_hline)(uint16_t x, uint16_t y, uint16_t len, uint32_t color){
     int r;
-    for (int i = 0; i < len; i++)
+    for (uint16_t i = 0; i < len; i++)
         if ((r = set_pixel(x + i, y, color))) return r;
     return SUCCESS;
 }
 
 int (vg_draw_rectangle)(uint16_t x, uint16_t y,uint16_t width, uint16_t height, uint32_t color)	{
     int r;
-    for (int i = 0; i < height; i++)
+    for (uint16_t i = 0; i < height; i++)
         if ((r = vg_draw_hline(x, y + i, width, color))) return r;
-
     return SUCCESS;
 }
